GitHub/arete_mr/requirements.txt
google-cloud-speech
openai
languagetool-python
google-cloud-language
google-cloud-storage
google-cloud-functions
pandas
numpy
flask
requests
spacy
